Extract into files
