+++ lib/arel/middleware/no_op_cache.rb
@@ -1,0 +1,9 @@
+module Arel
+  module Middleware
+    module NoOpCache
+      def self.get(key); end
+      def self.set(key, sql); end
+    end
+  end
+end
+
